alterando int_to_str para suportar uint64_t

+++ include/utils.h
@@ -2,6 +2,7 @@
 #define UTILS_H
 
 #include <stddef.h>
+#include <stdint.h>
 
 #define SECOND 1000
 
@@ -11,7 +12,7 @@ void get_input(char *input, size_t size);
 bool is_blank(const char *input);
 char *to_lower(const char *str);
 void clear_input_buffer();
-void int_to_str(int value, char *str, size_t size);
+void int_to_str(uint64_t value, char *str, size_t size);
 void float_to_str(float value, char *str, size_t size);
 
 #endif
